# users/users-service.js
import express from 'express';
import swaggerUi from 'swagger-ui-express';
import fs from 'node:fs';
import YAML from 'js-yaml';
import promBundle from 'express-prom-bundle';
import User from './src/models/User.js';
import GameRecord from './src/models/GameRecord.js';
import dotenv from 'dotenv';
import { createRequire } from 'module';
import { fileURLToPath } from 'url';
import connectDB from './src/database.js';  
import Hashing from './src/hashing.js';


//Carga las variables del .env
dotenv.config();

const app = express();
const port = 3000;

//conecta con mongoDB al arrancar el servidor
connectDB();

//Bloque condicional neceseario para los test al hacer deploy, borra los usuarios durante los test
//E2E, evitando errores de duplicado en la base de datos
if (process.env.NODE_ENV === 'test') {
  app.delete('/testing/deleteuser/:username', async (req, res) => {
    //Miramos el parámetro para evitar inyección en la query, por el aviso de SonarCloud
    const username = String(req.params.username);
    await User.deleteOne({ username: username });
    res.status(200).json({ message: 'User deleted' });
  });
}

//Añade metricas para Prometheus
const metricsMiddleware = promBundle({ includeMethod: true });
app.use(metricsMiddleware);

//Carga la documentacion de Swagger desde el archivo openapi.yaml
try {
  const swaggerDocument = YAML.load(fs.readFileSync('./openapi.yaml', 'utf8'));
  app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
} catch (e) {
  console.log(e);
}

//Este bloque permite que el fronted pueda hacer peticiones al backend
app.use((req, res, next) => {
  res.setHeader('Access-Control-Allow-Origin', '*');
  res.setHeader('Access-Control-Allow-Methods', 'GET,POST,OPTIONS');
  res.setHeader('Access-Control-Allow-Headers', 'Content-Type');
  if (req.method === 'OPTIONS') return res.sendStatus(204);
  next();
});

//Permite que el servido entienda la peticiones en formato JSON
app.use(express.json());


//ENDPOINT POST /createuser, recibe un username, lo guarda en mongoDB y responde con el mensaje de bienvenida
app.post('/createuser', async (req, res) => {
  const username = req.body.username ? String(req.body.username) : null;
  const password = req.body.password ? String(req.body.password) : null;

  try {
    //Si no hay username y/o password, devuelve error.
    if (!username || !password) {
      return res.status(400).json({ error: "Username and password are required" });
    }

    //Si hay username, crea el usuario, y lo guarda
    const hashedPassword = await Hashing.hashPassword(password);
    const newUser = new User({ username, password: hashedPassword });
    await newUser.save();

    //Mensaje de bienvenida
    res.status(201).json({
      message: `Hello ${username}!`,
      user: { username: newUser.username }
    });

  } catch (err) {
    if (err.code === 11000) {
      return res.status(409).json({ 
        error: `The username '${username}' is already taken. Please choose another one.` 
      });
    }
    //Si hay algun error como un usario que ya existe, responde con 400.
    res.status(400).json({ error: err.message });
  }
});

//Solo arranca el servidor si el archivo se ejecuta directamente, NO cuando lo importan los tests.
const __filename = fileURLToPath(import.meta.url);
if (process.argv[1] === __filename) {
  app.listen(port, () => {
    console.log(`User Service listening at http://localhost:${port}`)
  });
}

// ENDPOINT POST /login
app.post('/login', async (req, res) => {
  const username = req.body.username ? String(req.body.username) : null;
  const password = req.body.password ? String(req.body.password) : null;

  try {
    if (!username || !password) {
      return res.status(400).json({ error: "Username and password are required" });
    }
    //Buscamos al usuario en la base de datos
    const user = await User.findOne({ username });

    if (!user) {
      return res.status(401).json({ error: "User not found" });
    }

    // comparar la contraseña (usando bcrypt en el futuro)
    if (!(await Hashing.verifyPassword(user.password, password))) {
       return res.status(401).json({ error: "Invalid password" });
    }

    res.status(200).json({ 
      message: `Welcome back, ${username}!`, 
      user: { username: user.username } 
    });

  } catch (err) {
    console.error("Internal server error details:", err);
    res.status(500).json({ error: "Internal server error" });
  }
});

// ENDPOINT POST /savegame
// Guarda el resultado de una partida finalizada en el historial
// Body: { username, rival, resultado, size }
// resultado: '1' (gana el usuario logueado), '2' (pierde), 'X' (empate)
app.post('/savegame', async (req, res) => {
  const { username, rival, resultado, size } = req.body ?? {};

  if (!username || !rival || !resultado) {
    return res.status(400).json({ error: 'username, rival and resultado are required' });
  }

  if (!['1', '2'].includes(resultado)) {
    return res.status(400).json({ error: "resultado must be '1' or '2'" });
  }

  try {
    const record = new GameRecord({ username, rival, resultado, size});
    await record.save();
    res.status(201).json({ message: 'Game saved', record });
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

// ENDPOINT GET /history/:username
// Devuelve el historial de partidas de un usuario ordenado por fecha descendente
// Query params opcionales:
//   resultado:  '1' | '2' | 'X'
//   rival:      string (parcial, insensible a mayúsculas)
//   fechaDesde: string (ISO date: YYYY-MM-DD)
//   fechaHasta: string (ISO date: YYYY-MM-DD)
//   size:       number
app.get('/history/:username', async (req, res) => {
  const username = String(req.params.username);
  const { resultado, rival, fechaDesde, fechaHasta, size } = req.query;
 
  if (resultado && !['1', '2', 'X'].includes(resultado)) {
    return res.status(400).json({ error: "resultado must be '1', '2' or 'X'" });
  }
  if (size && isNaN(Number(size))) {
    return res.status(400).json({ error: "size must be a number" });
  }
 
  try {
    const filter = { username };
    if (resultado)  filter.resultado = resultado;
    if (rival)      filter.rival = { $regex: rival, $options: 'i' };
    if (size)       filter.size = Number(size);
 
    if (fechaDesde || fechaHasta) {
      filter.createdAt = {};
      if (fechaDesde) filter.createdAt.$gte = new Date(fechaDesde);
      if (fechaHasta) {
        const hasta = new Date(fechaHasta);
        hasta.setDate(hasta.getDate() + 1);
        filter.createdAt.$lt = hasta;
      }
    }
 
    const records = await GameRecord.find(filter)
      .sort({ createdAt: -1 })
      .lean();
    res.status(200).json({ username, history: records });
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

// ENDPOINT GET /stats/:username
app.get('/stats/:username', async (req, res) => {
  const username = String(req.params.username);

  try {
    const records = await GameRecord.find({ username }).sort({ createdAt: -1 }).lean();

    const total = records.length;
    if (total === 0) {
      return res.status(200).json({
        username, total: 0, wins: 0, losses: 0,
        winRate: 0, currentStreak: 0, bestStreak: 0,
        mostPlayedRival: null, rivalStats: {}
      });
    }

    const wins   = records.filter(r => r.resultado === '1').length;
    const losses = records.filter(r => r.resultado === '2').length;
    const winRate = Math.round((wins / total) * 1000) / 10;

    // Racha actual
    let currentStreak = 0;
    for (const r of records) {
      if (r.resultado === '1') currentStreak++;
      else break;
    }

    // Mejor racha histórica
    let bestStreak = 0;
    let streak = 0;
    for (const r of [...records].reverse()) {
      if (r.resultado === '1') { streak++; if (streak > bestStreak) bestStreak = streak; }
      else streak = 0;
    }

    // Estadísticas por rival
    const rivalMap = {};
    for (const r of records) {
      if (!rivalMap[r.rival]) rivalMap[r.rival] = { wins: 0, losses: 0, total: 0 };
      rivalMap[r.rival].total++;
      if (r.resultado === '1') rivalMap[r.rival].wins++;
      else                     rivalMap[r.rival].losses++;
    }

    const mostPlayedRival = Object.entries(rivalMap)
      .sort((a, b) => b[1].total - a[1].total)[0][0];

    res.status(200).json({
      username, total, wins, losses, winRate,
      currentStreak, bestStreak, mostPlayedRival, rivalStats: rivalMap
    });

  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

export default app;